delete only if any file except . and .. were selected

+++ lib/rfd/commands.rb
@@ -9,9 +9,11 @@ module Rfd
     end
 
     def d
-      FileUtils.mv selected_items.map(&:path), File.expand_path('~/.Trash/')
-      @row -= selected_items.count {|i| i.index <= @row}
-      ls
+      if selected_items.any?
+        FileUtils.mv selected_items.map(&:path), File.expand_path('~/.Trash/')
+        @row -= selected_items.count {|i| i.index <= @row}
+        ls
+      end
     end
 
     def f
@@ -67,9 +69,11 @@ module Rfd
     end
 
     def D
-      FileUtils.rm_rf selected_items.map(&:path)
-      @row -= selected_items.count {|i| i.index <= @row}
-      ls
+      if selected_items.any?
+        FileUtils.rm_rf selected_items.map(&:path)
+        @row -= selected_items.count {|i| i.index <= @row}
+        ls
+      end
     end
 
     def H
